refactor: move api request & response data classes to model dir
