+++ app/src/main/java/ru/practicum/android/diploma/search/data/mapper/VacancyDetailsMapper.kt
@@ -1,7 +1,7 @@
 package ru.practicum.android.diploma.search.data.mapper
 
 import ru.practicum.android.diploma.common.utils.FormatStrings
-import ru.practicum.android.diploma.search.data.source.remote.HHApiResponse
+import ru.practicum.android.diploma.search.data.model.HHApiResponse
 import ru.practicum.android.diploma.vacancy.domain.model.VacancyDetails
 
 fun HHApiResponse.VacancyDetails.toVacancyDetails() = VacancyDetails(
