+++ app/src/main/java/ru/practicum/android/diploma/search/data/repository/VacanciesGetterImpl.kt
@@ -4,8 +4,8 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import ru.practicum.android.diploma.common.domain.model.Vacancy
 import ru.practicum.android.diploma.filter.domain.model.Filter
-import ru.practicum.android.diploma.search.data.source.remote.HHApiRequest
-import ru.practicum.android.diploma.search.data.source.remote.HHApiResponse
+import ru.practicum.android.diploma.search.data.model.HHApiRequest
+import ru.practicum.android.diploma.search.data.model.HHApiResponse
 import ru.practicum.android.diploma.search.data.source.remote.NetworkClient
 import ru.practicum.android.diploma.search.data.source.remote.RetrofitClient.Companion.BAD_REQUEST
 import ru.practicum.android.diploma.search.data.source.remote.RetrofitClient.Companion.FORBIDDEN
