+++ app/src/main/java/ru/practicum/android/diploma/search/data/source/remote/HHApiService.kt
@@ -3,6 +3,7 @@ package ru.practicum.android.diploma.search.data.source.remote
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.QueryMap
+import ru.practicum.android.diploma.search.data.model.HHApiResponse
 
 interface HHApiService {
