+++ app/src/main/java/ru/practicum/android/diploma/search/data/source/remote/NetworkClient.kt
@@ -1,5 +1,7 @@
 package ru.practicum.android.diploma.search.data.source.remote
 
+import ru.practicum.android.diploma.search.data.model.HHApiResponse
+
 interface NetworkClient {
     suspend fun doRequest(request: Any): HHApiResponse
 }
